+++ CHANGELOG.md
@@ -8,10 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Removed SMS Search API
 - Deprecated Redact client
 - Use `vonage-jwt-jdk:1.0.2` library instead of `nexmo-jwt-jdk:1.0.1`
+- Ensure `User-Agent` is set in request headers
 - Allow alphanumeric characters for SMS and MMS sender fields in Messages API
 - Fixed incorrect restrictions on `WhatsappTemplateRequest` 
   - Policy is now optional
-  - Default locale is `en`
+  - Default locale is now `en`
   - Locale is now an enum rather than String
   - `parameters` is now `List<String>`
 - Removed dependency on `commons-io` and `commons-lang3`